feat(gltf_auto_export): massive restructure, wip
