+++ tools/gltf_auto_export/bevy_dynamic.py
