+++ tools/gltf_auto_export/export_materials.py
@@ -1,10 +1,10 @@
-
-from .helpers_collections import set_active_collection
-from .helpers_export import export_gltf, generate_gltf_export_preferences
-from .helpers import make_cube, traverse_tree
-import bpy
 import os
+import bpy
 from pathlib import Path
+
+from .helpers_collections import (set_active_collection, traverse_tree)
+from .auto_export.export_gltf import (export_gltf, generate_gltf_export_preferences)
+from .auto_export.object_makers import make_cube
 
 # get materials per object, and injects the materialInfo component
 def get_materials(object):
